Merges of Jordan's working code on nylon
