+++ examples/etc/config.yaml
@@ -4,7 +4,6 @@ common:
     server: True
 
 server:
-    web_directory: /home/infoset/public_html
     data_directory: /opt/infoset/cache/data
     ingest_cache_directory: /opt/infoset/cache/ingest
     ingest_threads: 20
