+++ infoset/charts/__init__.py
@@ -1,0 +1,11 @@
+"""Infoset utilities package.
+
+This package's modules which perform important tasks within
+the project but are either not specific enough or not large
+enough to warrant their own package
+
+"""
+
+# Import classes
+from infoset.charts.timestamp import TimeStamp
+from infoset.charts.colorwheel import ColorWheel
