+++ infoset/utils/__init__.py
@@ -10,9 +10,6 @@ enough to warrant their own package
 from infoset.utils.jm_configuration import Config
 from infoset.utils.jm_configuration import ConfigAgent
 from infoset.utils.jm_configuration import ConfigSNMP
-from infoset.utils.xlate_snmp import Translator
 from infoset.utils.hidden import Directory
 from infoset.utils.hidden import File
 from infoset.utils.daemon import Daemon
-from infoset.utils.timestamp import TimeStamp
-from infoset.utils.colorwheel import ColorWheel
