--- infoset/web/README.md
@@ -1,7 +1,0 @@
-# About
-This module contains files and sub-modules for rendering web pages
-
-# Submodules
-The following submodules are located here:
-
-1. **snmp**: Classes that render web pages using snmp data
