--- infoset/web/__init__.py
@@ -1,1 +1,0 @@
-"""Infoset web package."""
