+++ setup.py
@@ -1,90 +1,239 @@
 #!/usr/bin/env python3
-"""Infoset setup script.
+"""Infoset ORM classes.
 
-Run this script when setting up for the first time
+Manages connection pooling among other things.
 
 """
 
-import os
-import sys
-from setuptools import setup, find_packages
+# Main python libraries
+import socket
+from pathlib import Path
+
+# Pip3 libraries
+from sqlalchemy import create_engine
+
+# Infoset libraries
+try:
+    from infoset.utils import log
+except:
+    print('You need to set your PYTHONPATH to include the infoset library')
+    sys.exit(2)
+from infoset.utils import jm_configuration
+from infoset.utils import jm_general
+from infoset.metadata import metadata
+import infoset.utils
+from infoset.db.db_orm import BASE, Agent, Department, Host, BillType
+from infoset.db.db_orm import Configuration, HostAgent
+from infoset.db import DBURL
+from infoset.db import db_agent
+from infoset.db import db_configuration
+from infoset.db import db_billtype
+from infoset.db import db_department
+from infoset.db import db_host
+from infoset.db import db_hostagent
+from infoset.db import db
+from infoset.agents import agent
 
 
-# Define key files
-SETUP_SCRIPT_DIRECTORY = os.path.abspath(os.path.dirname(__file__))
+def insert_department():
+    """Insert first department in the database.
 
-README_FILE = open(
-    os.path.join(SETUP_SCRIPT_DIRECTORY, 'README.md')).read()
-NEWS_FILE = open(
-    os.path.join(SETUP_SCRIPT_DIRECTORY, 'NEWS.txt')).read()
+    Args:
+        None
 
-# Define other parameters
-VERSION = '0.0.0.0'
-PYPI = [
-    # List your project dependencies here.
-    # For more details, see:
-    # http://packages.python.org/distribute/setuptools.html#declaring-dependencies
-    'astroid >= 1.4.8',
-    'click >= 6.6',
-    'colorama >= 0.3.7',
-    'Flask >= 0.11.1',
-    'Flask-RESTful >= 0.3.5',
-    'funcsigs >= 1.0.2',
-    'isort >= 4.2.5',
-    'itsdangerous >= 0.24',
-    'Jinja2 >= 2.8',
-    'lazy-object-proxy >= 1.2.2',
-    'MarkupSafe >= 0.23',
-    'mccabe >= 0.5.2',
-    'mock >= 2.0.0',
-    'nose >= 1.3.7',
-    'numpy >= 1.11.1',
-    'pbr >= 1.10.0',
-    'pep257 >= 0.7.0',
-    'pep8 >= 1.7.0',
-    'ply >= 3.8',
-    'psutil >= 4.3.0',
-    'pyasn1 >= 0.1.9',
-    'pycrypto >= 2.6.1',
-    'pylint >= 1.6.4',
-    'PyMySQL >= 0.7.6',
-    'pysmi >= 0.0.7',
-    'pysnmp >= 4.3.2',
-    'python-dateutil >= 2.5.3',
-    'pytz >= 2016.6.1',
-    'PyYAML >= 3.11',
-    'requests >= 2.11.0',
-    'six >= 1.10.0',
-    'SQLAlchemy >= 1.0.14',
-    'Werkzeug >= 0.11.10',
-    'wrapt >= 1.10.8'
-]
+    Returns:
+        None
 
-setup(
-    name='Infoset',
-    version=VERSION,
-    description=(
-        'System performance charting and network topology system'),
-    long_description=README_FILE + '\n\n' + NEWS_FILE,
-    classifiers=[
-        # Get strings from
-        # http://pypi.python.org/pypi?%3Aaction=list_classifiers
-        'Environment :: Console',
-        'Programming Language :: Python :: 3'
-    ],
-    keywords='infoset snmp switchmap network map',
-    author='UWIComputingSociety',
-    author_email='',
-    url='',
-    license='',
-    packages=find_packages(
-        exclude=['*.tests', '*.tests.*', 'tests.*', 'tests']),
-    scripts=['setup/install/install.py'],
-    include_package_data=True,
-    zip_safe=False,
-    install_requires=PYPI,
-    entry_points={
-        'console_scripts':
-        ['infoset=infoset.toolbox:main']
-    }
-)
+    """
+    # Initialize key variables
+    if db_department.idx_exists(1) is False:
+        record = Department(
+            code=jm_general.encode('_SYSTEM_RESERVED_'),
+            name=jm_general.encode('_SYSTEM_RESERVED_'))
+        database = db.Database()
+        database.add(record, 1102)
+
+
+def insert_billtype():
+    """Insert first billtype in the database.
+
+    Args:
+        None
+
+    Returns:
+        None
+
+    """
+    # Initialize key variables
+    if db_billtype.idx_exists(1) is False:
+        record = BillType(
+            code=jm_general.encode('_SYSTEM_RESERVED_'),
+            name=jm_general.encode('_SYSTEM_RESERVED_'))
+        database = db.Database()
+        database.add(record, 1104)
+
+
+def insert_agent_host():
+    """Insert first agent and host in the database.
+
+    Args:
+        None
+
+    Returns:
+        None
+
+    """
+    # Initialize key variables
+    idx_agent = 1
+    idx_host = 1
+    agent_name = '_infoset'
+
+    # Add agent
+    if db_agent.idx_exists(idx_agent) is False:
+        record = Agent(
+            id=jm_general.encode('_SYSTEM_RESERVED_'),
+            name=jm_general.encode(agent_name))
+        database = db.Database()
+        database.add(record, 1109)
+
+        # Generate a UID
+        uid = agent.get_uid(agent_name)
+        database = db.Database()
+        session = database.session()
+        record = session.query(Agent).filter(Agent.idx == idx_agent).one()
+        record.id = jm_general.encode(uid)
+        database.commit(session, 1073)
+
+    # Add host
+    if db_host.idx_exists(idx_host) is False:
+        record = Host(
+            description=jm_general.encode('Infoset Server'),
+            hostname=jm_general.encode(socket.getfqdn()))
+        database = db.Database()
+        database.add(record, 1106)
+
+    # Add to Agent / Host table
+    if db_hostagent.host_agent_exists(idx_host, idx_agent) is False:
+        record = HostAgent(idx_host=idx_host, idx_agent=idx_agent)
+        database = db.Database()
+        database.add(record, 1107)
+
+
+def insert_config():
+    """Insert first config in the database.
+
+    Args:
+        None
+
+    Returns:
+        None
+
+    """
+    # Initialize key variables
+    key_values = [('version', '0.0.0.0')]
+
+    # Cycle through all the key value pairs
+    for item in key_values:
+        key = item[0]
+        value = item[1]
+
+        # Check if value exists and insert if not
+        if db_configuration.config_key_exists(key) is False:
+            record = Configuration(
+                config_key=jm_general.encode(key),
+                config_value=jm_general.encode(value))
+            database = db.Database()
+            database.add(record, 1108)
+
+
+def server_setup():
+    """Setup server.
+
+    Args:
+        None
+
+    Returns:
+        None
+
+    """
+    # Initialize key variables
+    use_mysql = True
+    pool_size = 25
+    max_overflow = 25
+
+    # Get configuration
+    config = jm_configuration.Config()
+
+    # Create DB connection pool
+    if use_mysql is True:
+        # Add MySQL to the pool
+        engine = create_engine(
+            DBURL, echo=True,
+            encoding='utf8',
+            max_overflow=max_overflow,
+            pool_size=pool_size, pool_recycle=3600)
+
+        # Try to create the database
+        print('Attempting to create database tables')
+        try:
+            sql_string = (
+                'ALTER DATABASE %s CHARACTER SET utf8mb4 '
+                'COLLATE utf8mb4_general_ci') % (config.db_name())
+            engine.execute(sql_string)
+        except:
+            log_message = (
+                'Cannot connect to database %s. '
+                'Verify database server is started. '
+                'Verify database is created. '
+                'Verify that the configured database authentication '
+                'is correct.') % (config.db_name())
+            log.log2die(1036, log_message)
+
+        # Apply schemas
+        print('Applying Schemas')
+        BASE.metadata.create_all(engine)
+
+        # Insert database entries
+        insert_agent_host()
+        insert_billtype()
+        insert_department()
+        insert_config()
+
+        # Try some additional statements
+        metadata.insert_oids()
+
+
+def main():
+    """Process agent data.
+
+    Args:
+        None
+
+    Returns:
+        None
+
+    """
+    # Get configuration
+    config = jm_configuration.Config()
+
+    # Run server setup if required
+    if config.server() is True:
+        server_setup()
+
+    # Install required PIP packages
+    print('Installing required pip3 packages')
+    pip3 = infoset.utils.jm_general.search_file('pip3')
+    if pip3 is None:
+        log_message = ('Cannot find python "pip3". Please install.')
+        log.log2die(1052, log_message)
+
+    utils_directory = infoset.utils.__path__[0]
+    requirements_file = ('%s/requirements.txt') % (
+        Path(utils_directory).parents[1])
+    script_name = (
+        'pip3 install --user --requirement %s') % (requirements_file)
+    infoset.utils.jm_general.run_script(script_name)
+
+
+if __name__ == '__main__':
+    main()
